+++ SignalServiceKit/Upload/UploadMetadata.swift
@@ -9,8 +9,8 @@ import Foundation
 public protocol UploadMetadata {
     var encryptedDataLength: UInt32 { get }
 
-    static var maxUploadSizeBytes: UInt { get }
-    static var maxPlaintextSizeBytes: UInt { get }
+    static var maxUploadSizeBytes: UInt64 { get }
+    static var maxPlaintextSizeBytes: UInt64 { get }
 }
 
 /// Includes extra info like digest for validation.
